remove type on global var

+++ src/VGElement.jl
@@ -5,7 +5,7 @@
 # - they have an id, used for reference in final VG
 # - they hold their dependencies to other elements in the tracking field
 
-elements_counter::Int = 1
+elements_counter = 1
 
 struct VGElement{T}
     __trie::VGTrie
